Fixed Homepage and started dynamic form limiting code

+++ db/migrate/20170814230831_create_requisitions.rb
@@ -5,7 +5,7 @@ class CreateRequisitions < ActiveRecord::Migration[5.1]
       t.integer :address_years
       t.integer :company_years
       t.string :marital_status
-      t.integer :requested_amount
+      t.bigint :requested_amount
       t.string :payment_terms
       t.string :bank
       t.text :comment
